Added several Subevents integral to AttackRoll.

+++ RPGLC/com/rpglc/core/RPGLClass.cs
@@ -50,11 +50,11 @@ public class RPGLClass : DatabaseContent {
     }
 
     public long? GetSubclassLevel() {
-        return GetInt("subclass_level");
+        return GetLong("subclass_level");
     }
 
     public RPGLClass SetSubclassLevel(long? subclassLevel) {
-        PutInt("subevent_level", subclassLevel);
+        PutLong("subevent_level", subclassLevel);
         return this;
     }
 
@@ -84,7 +84,7 @@ public class RPGLClass : DatabaseContent {
             _ = rpglObject.GetClasses().AddJsonObject(new JsonObject()
                 .PutString("name", GetName())
                 .PutString("id", GetDatapackId())
-                .PutInt("level", level)
+                .PutLong("level", level)
                 .PutJsonObject("additional_nested_classes", new())
             );
         } else {
@@ -92,7 +92,7 @@ public class RPGLClass : DatabaseContent {
             for (int i = 0; i < classes.Count(); i++) {
                 JsonObject classData = classes.GetJsonObject(i);
                 if (GetDatapackId() == classData.GetString("id")) {
-                    classData.PutInt("level", level);
+                    classData.PutLong("level", level);
                     break;
                 }
             }
